Part 1 of the insertion sort module with tests.
Weird excercise that I struggled on a bit.
In the end I got it working and had to write a test case to catch
cases where the lowest number is to be shifted all the way to array
position 0.

I still don't get why the excercise called for putting the currently
evaluated number in the array as a placeholder instead of the number
being evaluated.